+++ src/main/java/fuzs/betteranimationscollection/client/element/OinkyPigElement.java
@@ -1,14 +1,25 @@
 package fuzs.betteranimationscollection.client.element;
 
 import fuzs.betteranimationscollection.client.renderer.entity.model.OinkyPigModel;
+import fuzs.betteranimationscollection.mixin.client.accessor.ISaddleLayerAccessor;
+import net.minecraft.client.renderer.entity.layers.LayerRenderer;
 import net.minecraft.client.renderer.entity.layers.SaddleLayer;
 import net.minecraft.client.renderer.entity.model.EntityModel;
-import net.minecraft.client.renderer.entity.model.PigModel;
+import net.minecraft.entity.Entity;
+import net.minecraft.entity.IEquipable;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.passive.PigEntity;
+import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.SoundEvents;
 
-public class OinkyPigElement extends ModelElement {
+public class OinkyPigElement extends SoundModelElement {
+
+    public OinkyPigElement() {
+
+        super(PigEntity.class);
+        this.addDefaultSound(SoundEvents.PIG_AMBIENT);
+        this.addDefaultSound(new ResourceLocation("snowpig", "entity.snow_pig.ambient"));
+    }
 
     @Override
     public String[] getDescription() {
@@ -17,12 +28,16 @@ public class OinkyPigElement extends ModelElement {
                 "It only moves up and down ever so slightly, but it's there. Just a little bit more life for your livestock."};
     }
 
-    @SuppressWarnings("unchecked")
     @Override
     public void setupClient() {
 
-        this.addLayerTransformer(layerRenderer -> layerRenderer instanceof SaddleLayer, layerRenderer -> ((SaddleLayer<PigEntity, PigModel<PigEntity>>) layerRenderer).model = new OinkyPigModel<>(0.5F));
-        SoundDetectionElement.addNoisyEntity(PigEntity.class, SoundEvents.PIG_AMBIENT);
+        this.addLayerTransformer(layerRenderer -> layerRenderer instanceof SaddleLayer, this::setSaddleLayerModel);
+    }
+
+    @SuppressWarnings("unchecked")
+    private <T extends Entity & IEquipable, M extends EntityModel<T>> void setSaddleLayerModel(LayerRenderer<?, ?> layerRenderer) {
+
+        ((ISaddleLayerAccessor<T, M>) layerRenderer).setModel((M) new OinkyPigModel<T>(0.5F));
     }
 
     @Override